Extract shared RailML primitives into model/common.py
NS, _NS, _NSMAP, _Base, _make_id, Designator and Name were defined in
rollingstock.py and imported from there by infrastructure.py, creating a
dependency between two sibling sub-schema modules.

Moving them to model/common.py gives each sub-schema a clean, neutral
import source.  rollingstock.py re-exports everything via `noqa: F401` so
existing callers (make_railml_rollingstock.py, run_100trains.py,
rollingstock_ui.py) continue to work unchanged.
